# docs/swagger.yaml
basePath: /api/v1
definitions:
  books.bookRequest:
    properties:
      description:
        type: string
      image_url:
        type: string
      published_date:
        type: string
      title:
        type: string
    required:
    - description
    - published_date
    - title
    type: object
  model.Book:
    properties:
      book_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      image_url:
        type: string
      published_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:3080
info:
  contact:
    email: gmhafiz@gmail.com
    name: Hafiz Shafruddin
    url: http://www.gmhafiz.com/contact
  description: Go + Postgres + Chi Router + SqlBoiler starter kit for API development.
  license: {}
  title: Go8
  version: 0.2.0
paths:
  /books:
    get:
      consumes:
      - application/json
      description: Get all books. By default it gets first page with 10 items.
      parameters:
      - description: page number
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Book'
            type: array
      summary: Show all books
    post:
      consumes:
      - application/json
      description: Get a book with JSON payload
      parameters:
      - description: Book Request
        in: body
        name: Book
        required: true
        schema:
          $ref: '#/definitions/books.bookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Book'
      summary: Create a Book
  /books/{bookID}:
    delete:
      consumes:
      - application/json
      description: Delete a book by its id.
      parameters:
      - description: book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
        "500":
          description: Internal Server error
      summary: Delete a Book
    get:
      consumes:
      - application/json
      description: Get a book by its id.
      parameters:
      - description: book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Book'
      summary: Get a Book
    put:
      consumes:
      - application/json
      description: Update a book by its model.
      parameters:
      - description: Book Request
        in: body
        name: Book
        required: true
        schema:
          $ref: '#/definitions/books.bookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
        "500":
          description: Internal Server error
      summary: Update a Book
  /books/search:
    get:
      consumes:
      - application/json
      description: Search a book by its title.
      parameters:
      - description: search term
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Book'
            type: array
      summary: Search a Book
swagger: "2.0"